+++ CHANGES.md
@@ -1,5 +1,15 @@
 # Release History
 
+## 0.1.11
+* Added new Milestone API - Create a new Milestone
+* Added new Milestone API - Edit the Milestone
+* Added new Milestone API - Approve the Milestone
+* Added new Milestone API - Activate the Milestone
+* Added new Milestone API - Delete the Milestone
+* Added new Submission API - Submit for Approval
+* Added new Submission API - Approve the Submission
+* Added new Submission API - Reject the Submission
+
 ## 0.1.10
 * Added new call for Referenced User API